Add validation file in store
