+++ store/service.go
@@ -3,7 +3,6 @@ package store
 import (
 	"context"
 	"errors"
-	"fmt"
 
 	"github.com/heetch/regula"
 	"github.com/heetch/regula/rule"
@@ -79,21 +78,4 @@ type RulesetEvent struct {
 type RulesetEvents struct {
 	Events   []RulesetEvent
 	Revision string
-}
-
-// ValidationError gives informations about the reason of failed validation.
-type ValidationError struct {
-	Field  string
-	Value  string
-	Reason string
-}
-
-func (v *ValidationError) Error() string {
-	return fmt.Sprintf("invalid %s with value '%s': %s", v.Field, v.Value, v.Reason)
-}
-
-// IsValidationError indicates if the given error is a ValidationError pointer.
-func IsValidationError(err error) bool {
-	_, ok := err.(*ValidationError)
-	return ok
 }
