+++ store/validation.go
@@ -1,0 +1,20 @@
+package store
+
+import "fmt"
+
+// ValidationError gives informations about the reason of failed validation.
+type ValidationError struct {
+	Field  string
+	Value  string
+	Reason string
+}
+
+func (v *ValidationError) Error() string {
+	return fmt.Sprintf("invalid %s with value '%s': %s", v.Field, v.Value, v.Reason)
+}
+
+// IsValidationError indicates if the given error is a ValidationError pointer.
+func IsValidationError(err error) bool {
+	_, ok := err.(*ValidationError)
+	return ok
+}
